chore: rename normalize file
