+++ jcloud/normalize.py
